- viewClip sample updated
