+++ src/frontend-samples/view-clip-sample/ViewClip.scss
@@ -1,0 +1,20 @@
+/*---------------------------------------------------------------------------------------------
+* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+* See LICENSE.md in the project root for license terms and full copyright notice.
+*--------------------------------------------------------------------------------------------*/
+.sample-options {
+  color: white;
+  padding: 8px;
+}
+
+.sample-options-3col {
+  display: grid;
+  row-gap: 8px;
+  column-gap: 8px;
+  grid-template-columns: 1fr 2fr 2fr;
+  align-items: center;
+}
+
+.even-3col {
+  grid-template-columns: 1fr 1fr 1fr;
+}
